backend/requirements.txt
pymongo==4.11.2
python-dotenv==1.0.1
Flask==3.1.0
